backend/requirements.txt
fastapi==0.125.0
uvicorn[standard]==0.38.0
python-multipart==0.0.21
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.45
python-dotenv==1.2.1
flask==3.0.0
flask-cors==4.0.0
numpy==2.3.5
opencv-python==4.10.0.84
pillow==10.3.0
requests==2.32.3
gunicorn==23.0.0
